refactor(#1430): split api/runtime.py into sub-package
The 1839-LOC `src/scistudio/api/runtime.py` god-file is decomposed into
`src/scistudio/api/runtime/` — 8 sub-modules each <500 LOC, with the
full public import surface preserved.

* `__init__.py` defines `ApiRuntime` directly so griffe emits the
  canonical `scistudio.api.runtime.ApiRuntime` symbol fact + its
  `<method>` attribute facts. Method bodies live in sub-modules
  (`_projects`, `_workflows`, `_data`, `_runs`) and are bound onto
  `ApiRuntime` via class-body static assignment.
* Dataclasses (`KnownProject`, `DataRecord`, `WorkflowRun`) and
  `LogBroadcaster` are defined at the package level so griffe emits
  `scistudio.api.runtime.<Name>` facts directly.
* Private helpers (`_table_cache`, `_table_cache_lock`,
  `_read_preview_table_from_disk`, `_rmtree_force`, …) are re-exported
  so existing tests reach in via the public package symbol
  unchanged. `_preview_cache._get_preview_table` resolves the disk
  reader via `sys.modules['scistudio.api.runtime']` so the LRU-cache
  test's `monkeypatch.setattr(runtime_mod, ...)` keeps observing.
* `tests/api/test_runtime_import_surface.py` (new) pins the public
  surface: importable names, shared cache identity, monkeypatch
  visibility, and the 41-method roster on `ApiRuntime`.
* `scripts/check_god_files.py`: the `api/runtime.py` waiver is retired
  (file replaced by sub-package, every sub-module <750 LOC).
* `docs/adr/ADR-{012,038,039,044,045}.md` frontmatter: `governs.files`
  entries updated from `src/scistudio/api/runtime.py` to
  `src/scistudio/api/runtime/` (directory) so the closure +
  doc-drift audits resolve cleanly post-split. No body changes.

Pure structural refactor: no behavior changes, no API changes, no
schema changes. Every previously-public name is still importable via
`from scistudio.api.runtime import X` and via
`from scistudio.api import runtime` + attribute access.

Part of umbrella #1427.

Closes #1430

Gate-Record: .workflow/records/1430-api-runtime.json
Task-Kind: refactor
Issue: #1430
