+++ docs/adr/ADR-012.md
@@ -22,7 +22,7 @@ governs:
   files:
     - docs/adr/ADR-012.md
     - src/scistudio/engine/checkpoint.py
-    - src/scistudio/api/runtime.py
+    - src/scistudio/api/runtime/
     - tests/engine/test_checkpoint.py
   excludes: []
 tests:
