+++ docs/adr/ADR-038.md
@@ -26,7 +26,7 @@ governs:
     - docs/adr/ADR-038.md
     - src/scistudio/core/lineage/**
     - src/scistudio/engine/lineage_recorder.py
-    - src/scistudio/api/runtime.py
+    - src/scistudio/api/runtime/
   excludes: []
 tests:
   - tests/core/test_lineage_store_4table.py
