+++ docs/adr/ADR-039.md
@@ -31,7 +31,7 @@ governs:
     - docs/adr/ADR-039.md
     - src/scistudio/core/versioning/**
     - src/scistudio/api/routes/git.py
-    - src/scistudio/api/runtime.py
+    - src/scistudio/api/runtime/
     - frontend/src/components/Git/**
   excludes: []
 tests:
