+++ docs/adr/ADR-044.md
@@ -28,7 +28,7 @@ governs:
     - src/scistudio/workflow/definition.py
     - src/scistudio/workflow/schema.py
     - src/scistudio/workflow/validator.py
-    - src/scistudio/api/runtime.py
+    - src/scistudio/api/runtime/
     - src/scistudio/core/lineage/recorder.py
     - docs/architecture/ARCHITECTURE.md
     - docs/specs/adr-044-subworkflow-block.md
