+++ docs/adr/ADR-045.md
@@ -29,7 +29,7 @@ governs:
     - src/scistudio/api/routes/workflows.py
     - src/scistudio/api/routes/projects.py
     - src/scistudio/api/routes/git.py
-    - src/scistudio/api/runtime.py
+    - src/scistudio/api/runtime/
     - src/scistudio/engine/events.py
     - frontend/src/hooks/useWebSocket.ts
     - frontend/src/store/tabSlice.ts
